refactor : divided helper to PdfHelpers
